Split API-only UI E2E coverage
