+++ tests/e2e/ui/multi-repo-flow.spec.ts
@@ -2,15 +2,9 @@
  * Multi-repo UI flow (Phase 4b) — acceptance criterion 20.
  *
  * Registers a multi-repo project (two real git fixtures + one data-only repo)
- * via the API, drives the Settings → Components UI, asserts the workflows
- * panel shows resolved (component, command) pairs, exercises the
- * multi-repo indicator on the goal-creation form, creates a goal and
- * checks per-repo worktrees on disk, then archives the goal and asserts
- * cleanup. Persistence-across-reload is verified via a hard reload.
- *
- * The per-component setup invocation trace is verified opportunistically:
- * if the running server has wired runComponentSetups (Phase 4a), the
- * record file picks up entries. Phase 4b alone doesn't depend on that.
+ * via the API, drives the Settings → Components UI, and verifies component
+ * rendering/editing/deletion browser flows. API/data-path coverage lives in
+ * tests/e2e/multi-repo-flow-api.spec.ts.
  */
 import { test, expect } from "../gateway-harness.js";
 import { apiFetch, readE2EToken, base } from "../e2e-setup.js";
@@ -172,35 +166,7 @@ test.describe("multi-repo flow (UI)", () => {
 		}
 	});
 
-	// NOTE: original test fills a `worktree-root-input` testid in the
-	// Components tab and clicks save. That input does not exist in the
-	// Components tab — the only worktree_root editor in production lives
-	// in the General settings tab (settings-page.ts L646), and it is not a
-	// per-project structured field exposed via
-	// `/api/projects/:id/structured`. Until a per-project worktree_root UI
-	// ships in the Components tab (or the structured endpoint adds it), we
-	// validate the underlying data path via the API directly: the
-	// structured endpoint must return `worktree_root` (possibly empty) and
-	// must accept a structured PUT to set it.
-	test("Settings → Components: worktree_root structured endpoint round-trips", async () => {
-		test.setTimeout(60_000);
-		const project = await registerMultiRepoProject();
-		try {
-			const customRoot = path.join(os.tmpdir(), `bobbit-wt-${Date.now()}`);
-			const put = await apiFetch(`/api/projects/${project.id}/config`, {
-				method: "PUT",
-				body: JSON.stringify({ worktree_root: customRoot }),
-			});
-			expect(put.status).toBeLessThan(300);
-
-			const res = await apiFetch(`/api/projects/${project.id}/structured`);
-			const data = await res.json();
-			expect(data.worktree_root).toBe(customRoot);
-		} finally {
-			await apiFetch(`/api/projects/${project.id}`, { method: "DELETE" }).catch(() => {});
-			project.cleanup();
-		}
-	});
+	// Structured worktree_root endpoint coverage lives in tests/e2e/multi-repo-flow-api.spec.ts.
 
 	// Re-scan-from-Settings flow was replaced with "Open Project Assistant";
 	// repo scanning now happens in the assistant's interactive flow rather than
@@ -234,101 +200,4 @@ test.describe("multi-repo flow (UI)", () => {
 		}
 	});
 
-	test("multi-repo goal: per-repo worktrees on disk, then archive cleanup", async () => {
-		test.setTimeout(120_000);
-		const project = await registerMultiRepoProject();
-
-		try {
-			// Drive the goal-creation entirely via the API so the test stays
-			// stable across UI flow changes; the UI affordance (multi-repo
-			// indicator) is asserted in a separate test below.
-			//
-			// `cwd` must be one of the configured repos so isGitRepo() returns
-			// true and `goal.repoPath` is set; createWorktreeSet then runs across
-			// every component.
-			const goalRes = await apiFetch("/api/goals", {
-				method: "POST",
-				body: JSON.stringify({
-					projectId: project.id,
-					title: "Multi-repo goal",
-					spec: "Spec",
-					workflowId: "simple",
-					cwd: path.join(project.rootPath, "api"),
-					autoStartTeam: false,
-					team: false,
-				}),
-			});
-			expect(goalRes.status).toBeLessThan(300);
-			const goal = await goalRes.json();
-
-			// Wait for setupStatus to settle — success or error. Phase 4a wiring
-			// for multi-repo goal worktrees is in progress; if the server didn't
-			// produce `repoWorktrees`, we just confirm the goal was created and
-			// move on (the API-level test in tests/e2e/multi-repo-goal.spec.ts
-			// will tighten once Phase 4a lands).
-			let goalRecord: any = goal;
-			for (let i = 0; i < 60; i++) {
-				const r = await apiFetch(`/api/goals/${goal.id}`);
-				goalRecord = await r.json();
-				if (goalRecord?.setupStatus === "ready" || goalRecord?.setupStatus === "error") break;
-				await new Promise(res => setTimeout(res, 500));
-			}
-			expect(goalRecord.id).toBeTruthy();
-
-			if (goalRecord.repoWorktrees && Object.keys(goalRecord.repoWorktrees).length > 1) {
-				// Phase 4a has wired multi-repo goal worktrees — assert per-repo paths exist.
-				expect(Object.keys(goalRecord.repoWorktrees)).toEqual(expect.arrayContaining(["api", "web", "shared"]));
-				for (const [, wtPath] of Object.entries(goalRecord.repoWorktrees as Record<string, string>)) {
-					expect(fs.existsSync(wtPath as string)).toBe(true);
-				}
-
-				// Archive → cleanup. Allow up to 15s for async teardown.
-				await apiFetch(`/api/goals/${goal.id}`, { method: "DELETE" });
-				let allGone = false;
-				for (let i = 0; i < 30; i++) {
-					allGone = Object.values(goalRecord.repoWorktrees as Record<string, string>)
-						.every(wtPath => !fs.existsSync(wtPath as string));
-					if (allGone) break;
-					await new Promise(res => setTimeout(res, 500));
-				}
-				expect(allGone).toBe(true);
-			} else {
-				// Pre-Phase-4a single-repo fallback: just verify a worktree was set up.
-				expect(goalRecord.worktreePath || goalRecord.cwd).toBeTruthy();
-				await apiFetch(`/api/goals/${goal.id}`, { method: "DELETE" });
-			}
-		} finally {
-			await apiFetch(`/api/projects/${project.id}`, { method: "DELETE" }).catch(() => {});
-			project.cleanup();
-		}
-	});
-
-	test("multi-repo data path: structured endpoint surfaces a >1 repo count for the goal-form indicator", async () => {
-		test.setTimeout(60_000);
-		const project = await registerMultiRepoProject();
-
-		try {
-			// The goal-form's multi-repo indicator only renders when its
-			// dependent state (previewProjectId, previewTitle, etc.) is hot
-			// in the running session — driving that via the mock-agent flow
-			// adds incidental fragility. The actual indicator template is
-			// guarded by `componentSummary?.multiRepo` which derives from
-			// `/api/projects/:id/structured`. This test asserts that data
-			// path: register the project, GET /structured, and verify a
-			// >1 repo count is reported.
-			const res = await apiFetch(`/api/projects/${project.id}/structured`);
-			expect(res.status).toBe(200);
-			const data = await res.json();
-			expect(Array.isArray(data?.components)).toBe(true);
-			const repos = new Set((data.components as Array<{ repo?: string }>).map(c => c.repo || "."));
-			expect(repos.size).toBeGreaterThan(1);
-			// Workflows panel + components rendering for the same project is
-			// already verified in the first test in this file; covering the
-			// goal-form indicator in the same suite would duplicate that
-			// browser-render concern without adding signal.
-		} finally {
-			await apiFetch(`/api/projects/${project.id}`, { method: "DELETE" }).catch(() => {});
-			project.cleanup();
-		}
-	});
 });
